+++ js/change-state.js
@@ -1,7 +1,7 @@
-import renderArtist from "./templates/artist-view-screen";
-import renderGenre from "./templates/genre-view-screen";
-import renderState from "./templates/change-state-game-view-screen";
-import renderSuccess from "./templates/succes-view-screen";
+import renderArtist from "./screens/artist-view-screen";
+import renderGenre from "./screens/genre-view-screen";
+import renderState from "./screens/change-state-game-view-screen";
+import renderSuccess from "./screens/succes-view-screen";
 import {game} from "./data/models/game";
 
 const changeView = (state) => {
