+++ js/screens/artist-view-screen.js
