+++ js/screens/change-state-game-view-screen.js
