+++ js/screens/welcome-view-screen.js
